Allow slice patterns for destructuring arrays & slices
This PR changes the `ProjectionElem::Index` variant so that it takes
an `Operand` instead of a `VarId`. It then changes the
`index_to_function_calls` pass to create a local when needed.

I'm using `Operand` as a way of encoding a `Either<VarId,
ConstantExpr>`, I'm not sure that's entirely satisfactory. The hax ->
ULLBC pass will never produce the variant `Operand::Move`, whence the
`panic!` I introduced in `index_to_function_calls`.